FIX: do not bind IOCs to 127.0.0.1
Because you can not do broadcast addressing to localhost, if we have
multiple IOCs all bound to localhost you have to let them bind to all
of the interfaces they can see (localhost + tap0).  The default
settings of the CA clients will also see the tap0 interface and send
the search to its broadcast address (10.0.2.255) which makes all of
the IOCs see the search request and be able to service it if
needed (as opposed to just one random IOC seeing it).

This is because for UDP sockets on the server side when using
SO_REUSEADDR and SO_REUSEPORT does multi-cast to all of the bound
servers for broadcast address but load-balancing from uni-cast.